moving Main container to separate directory
