+++ client/src/Root.js
@@ -2,12 +2,12 @@ import React from 'react';
 import './Root.scss';
 import { BrowserRouter as Router, Route } from 'react-router-dom';
 
-import Main from './containers/Main';
+// import Main from './containers/Main/Main';
 // import Garage from './containers/Garage/Garage';
 // import Game from './containers/Play/Play';
 import Shop from './containers/Shop';
 
-import { Garage, Game } from './containers';
+import { Garage, Game, Main } from './containers';
 
 function Root() {
   return (
