+++ client/src/containers/Main/Main.js
@@ -1,12 +1,12 @@
 import React, { Component } from 'react';
 // import axios from 'axios';
-import CarSlider from './../components/CarSlider/CarSlider';
-import SocialMedia from './../components/SocialMedia/SocialMedia';
-import PlayNow from './../components/PlayNow/PlayNow';
+import CarSlider from './../../components/CarSlider/CarSlider';
+import SocialMedia from './../../components/SocialMedia/SocialMedia';
+import PlayNow from './../../components/PlayNow/PlayNow';
 
-import tokenValidationHelper from '../lib/tokenValidationHelper';
+import tokenValidationHelper from './../../lib/tokenValidationHelper';
 
-import Header from './../components/Header/Header';
+import Header from './../../components/Header/Header';
 
 export default class MainPage extends Component {
   constructor() {
