+++ client/src/containers/index.js
@@ -1,2 +1,3 @@
 export { default as Game } from './Play/Play';
 export { default as Garage } from './Garage/Garage';
+export { default as Main } from './Main/Main';
